# oas3_allyoucanorder.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: "AllYouCanOrder"
  description: API for managing restaurant orders.
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Localhost
paths:
...
  /tavoliRisto:
    get:
      description: >-
          Questa post serve per inserire l'elemento selezionato in ristorante.menu nell'array tavolo.carrello viene chiamata in menu.js
      summary: Mostra tavoli
      parameters:
        - name: x-access-token
          in: header
          description: an authorization header
          required: true
          schema:
            type: string
          example: token.esempio.prova
      responses:
        '201':
          description: 'Lista di tavoli'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    self:
                      type: string
                      description: 'Link al tavolo'
                    id:
                      type: string
                      description: 'Id del tavolo'
                    nome:
                      type: string
                      description: 'Nome del tavolo'
                    chiamato:
                      type: boolean
                      description: 'Indica se il cameriere è stato chiamato'
              example:
                  - self: '/api/v1/tavoliRisto/629a4cb56815e55554e7d7ae'
                    id: '629a4cb56815e55554e7d7ae'
                    nome: 'Tavolo 01'
                    chiamato: false
                  - self: '/api/v1/tavoliRisto/629715ffe86e996e2f47ff5c'
                    id: '629715ffe86e996e2f47ff5c'
                    nome: 'Tavolo 02'
                    chiamato: true
        '404':
          description: 'Ristorante non trovato'
        '404':
          description: 'Tavoli non trovati'
...
  /tavoliRisto/rispondiChiamata:
    post:
      description: >-
        Questa post serve per settare tavolo.chiamato=false se il cameriere preme il pulsante "rispondiChiamata" presente nel file scriptC.js
      summary: Rispondi alla chiamata del tavolo
      requestBody:
        content:
            application/json:
                schema:
                  type: object
                  required:
                    - id
                  properties:
                    id:
                      type: string
                      description: 'Id del Tavolo'
                example:
                    id: '629a4cb56815e55554e7d7ae'
      responses:
        '201':
          description: 'Chiamata impostata a false'
...